+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Upcoming
 
 - Async beatmap parsing
+- Hide various parsing related types further inwards, i.e. `rosu_pp::parse::some_type` instead of `rosu_pp::some_type`
+  - Affected types: `DifficultyPoint`, `HitObject`, `Pos2`, `TimingPoint`, `HitObjectKind`, `PathType`, `HitSound`
 
 # v0.1.1